Added parent object and inheriting objects
Tracker is the generic object. Any tracking object inherits from it. I
gave a description in the file otherTracker.py for what capabilities it
has and where to put things. Look at the find() method in circleTracker
for an idea of what you need to return and what else needs to be done
once the object is found. Let me know if the new method does not
involve a radius data point or something similar because I will need to
tweak the parent class a little bit to accommodate that.